feat(bar-chart-card): move sample data to an external file
